Save Data / Update Data / History
Saving the game was adding more sessions every time the user interacted. Changed this so it only adds a session when the player first interacts, and afterward updates that session. Also added a History mechanic, but it still needs to be hooked up to the rest of the system as well as populated with more events.